Move TestingTrait into Util; fix dup testInit(); only create symlinks for test env when needed
